pyproject.toml
[project]
name = "gait"
version = "0.1.0"
description = "Local ML models via OpenAI-compatible REST APIs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "python-multipart>=0.0.18",
    "pydub>=0.25",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "pytest-httpx>=0.35",
    "openai>=1.60",
]
scripts = [
    "deepfilternet>=0.5",
    "torch>=2,<2.6",
    "torchaudio>=2,<2.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gateway"]

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["."]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "B008",  # function calls in defaults (FastAPI's File/Form/Depends pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
